pyproject.toml
[tool.poetry]
name = "lcsvendingmachinemanagement"
version = "0.1.0"
description = ""
authors = ["repl.it user <replituser@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
mysql-connector-python = "^8.0.23"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
